+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+* Added support for loading previous versions of the configuration file, while they are
+  still supported.
+
 ## 16.6.0 - 2026-02-12
 
 ### Added